[TESTUTILS] Add test utility library and include XState helpers
