+++ modules/rostests/CMakeLists.txt
@@ -25,6 +25,7 @@ endif()
 #add_subdirectory(regtests)
 add_subdirectory(rosautotest)
 add_subdirectory(tests)
+add_subdirectory(testutils)
 add_subdirectory(unittests)
 add_subdirectory(win32)
 add_subdirectory(winetestlib)
