+++ modules/rostests/apitests/umkm/XStateConfig.cpp
@@ -9,167 +9,7 @@
 #include <windows.h>
 #include <versionhelpers.h>
 #include <x86x64/Cpuid.h>
-
-// These are not officially documented
-#define XSTATE_PKRU                         9
-#define XSTATE_HDC                          13
-#define XSTATE_UINTR                        14
-#define XSTATE_LBR                          15
-#define XSTATE_MASK_PKRU                    (1LL << (XSTATE_PKRU))
-#define XSTATE_MASK_HDC                     (1LL << (XSTATE_HDC))
-#define XSTATE_MASK_UINTR                   (1LL << (XSTATE_UINTR))
-#define XSTATE_MASK_LBR                     (1LL << (XSTATE_LBR))
-
-#define XSTATE_MASK_SUPERVISOR \
-    (XSTATE_MASK_IPT | \
-     XSTATE_MASK_PASID | \
-     XSTATE_MASK_CET_U | \
-     XSTATE_MASK_CET_S | \
-     XSTATE_MASK_HDC | \
-     XSTATE_MASK_UINTR | \
-     XSTATE_MASK_LBR)
-
-template<ULONG NtDdiVersion>
-struct TXSTATE_CONFIGURATION;
-
-template<>
-struct TXSTATE_CONFIGURATION<NTDDI_WIN7>
-{
-    ULONGLONG EnabledFeatures;                                              //0x0
-    ULONG Size;                                                             //0x8
-    ULONG OptimizedSave:1;                                                  //0xc
-    struct _XSTATE_FEATURE Features[64];                                    //0x10
-};
-
-template<>
-struct TXSTATE_CONFIGURATION<NTDDI_WIN8>
-{
-    ULONGLONG EnabledFeatures;                                              //0x0
-    ULONGLONG EnabledVolatileFeatures;                                      //0x8
-    ULONG Size;                                                             //0x10
-    union
-    {
-        ULONG ControlFlags;                                                 //0x14
-        struct
-        {
-            ULONG OptimizedSave:1;                                          //0x14
-            ULONG CompactionEnabled:1;                                      //0x14
-            ULONG ExtendedFeatureDisable:1;                                 //0x14
-        };
-    };
-    struct _XSTATE_FEATURE Features[64];                                    //0x18
-};
-
-template<>
-struct TXSTATE_CONFIGURATION<NTDDI_WIN10> : TXSTATE_CONFIGURATION<NTDDI_WIN8>
-{
-    ULONGLONG EnabledSupervisorFeatures;                                    //0x218
-    ULONGLONG AlignedFeatures;                                              //0x220
-    ULONG AllFeatureSize;                                                   //0x228
-    ULONG AllFeatures[64];                                                  //0x22c
-};
-
-template<>
-struct TXSTATE_CONFIGURATION<NTDDI_WIN10_RS5> : TXSTATE_CONFIGURATION<NTDDI_WIN10>
-{
-    ULONGLONG EnabledUserVisibleSupervisorFeatures;                         //0x330
-};
-
-template<>
-struct TXSTATE_CONFIGURATION<NTDDI_WIN11> : TXSTATE_CONFIGURATION<NTDDI_WIN10_RS5>
-{
-    ULONGLONG ExtendedFeatureDisableFeatures;                               //0x338
-    ULONG AllNonLargeFeatureSize;                                           //0x340
-    ULONG Spare;                                                            //0x344
-};
-
-ULONG GetXStateNtDdiVersion(void)
-{
-    if (IsReactOS())
-    {
-        return NTDDI_WIN11;
-    }
-
-    /* Get the NTDDI_VERSION from the PEB fields */
-    PPEB Peb = NtCurrentPeb();
-    ULONG WinVersion = (Peb->OSMajorVersion << 8) | Peb->OSMinorVersion;
-    switch (WinVersion)
-    {
-        case _WIN32_WINNT_WIN7:
-            return NTDDI_WIN7;
-
-        case _WIN32_WINNT_WIN8:
-            return NTDDI_WIN8;
-
-        case _WIN32_WINNT_WINBLUE:
-            return NTDDI_WIN8; // same as Win8
-
-        case _WIN32_WINNT_WIN10:
-        {
-            switch (Peb->OSBuildNumber)
-            {
-                case 10240: // 10.0.10240 / 1507 / Threshold 1
-                case 10586: // 10.0.10586 / 1511 / Threshold 2
-                case 14393: // 10.0.14393 / 1607 / Redstone 1
-                case 15063: // 10.0.15063 / 1703 / Redstone 2
-                case 16299: // 10.0.16299 / 1709 / Redstone 3
-                case 17134: // 10.0.17134 / 1803 / Redstone 4
-                    return NTDDI_WIN10;
-                case 17763: // 10.0.17763 / 1809 / Redstone 5
-                case 18362: // 10.0.18362 / 1903 / 19H1 "Titanium"
-                case 18363: // 10.0.18363 / Vanadium
-                case 19041: // 10.0.19041 / 2004 / Vibranium R1
-                case 19042: // 10.0.19042 / 20H2 / Vibranium R2 aka Manganese
-                case 19043: // 10.0.19043 / 21H1 / Vibranium R3 aka Ferrum
-                case 19044: // 10.0.19044 / 21H2 / Vibranium R4 aka Cobalt
-                case 19045: // 10.0.19045 / 22H2 / Vibranium R5
-                    return NTDDI_WIN10_RS5;
-
-                // Win 11
-                case 22000: // Cobalt
-                case 22621: // 22H2 Nickel R1
-                case 22631: // 23H2 Nickel R2
-                case 26100: // 24H2 Germanium
-                    return NTDDI_WIN11;
-
-                default:
-                    trace("Unknown Windows 10 build number: %d\n", Peb->OSBuildNumber);
-                    return 0;
-            }
-        }
-
-        default:
-            trace("UnsuUnknown Windows version: 0x%lX\n", WinVersion);
-            return 0;
-    }
-
-    return 0;
-}
-
-template<ULONG NtDdiVersion>
-SIZE_T GetXStateOffset(void)
-{
-    if (IsReactOS())
-    {
-        return FIELD_OFFSET(KUSER_SHARED_DATA, XState); // ReactOS
-    }
-    if (NtDdiVersion < NTDDI_WIN8)
-    {
-        return 0x3e0; // Win 7
-    }
-    else
-    {
-        return 0x3d8; // Win 8 - Win 11
-    }
-}
-
-template<ULONG NtDdiVersion>
-TXSTATE_CONFIGURATION<NtDdiVersion>* GetOsXState(void)
-{
-    SIZE_T Offset = GetXStateOffset<NtDdiVersion>();
-    PVOID Pointer = (PVOID)((ULONG_PTR)SharedUserData + Offset);
-    return (TXSTATE_CONFIGURATION<NtDdiVersion>*)Pointer;
-}
+#include <XStateHelpers.h>
 
 void GetExpectedXStateConfig(TXSTATE_CONFIGURATION<NTDDI_WIN11>* XStateConfig)
 {
